pyproject.toml
[project]
name = "resume-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
]
